+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.2.4
 - Fixed a bug where `pyflow script` was broken
+- Fixed a bug where `pyflow init` was broken
+- Fixed parsing `Pipfile`
 - `pyflow switch` now sets up the environment/dependencies with the new version
 - Fixed a bug where `bsd` specified as an OS on Pypi would cause a crash